move camera function to private-room.component.ts
