+++ app/client/src/modules/scenes/private-room/private-room.component.ts
@@ -2,6 +2,7 @@ import {
   container,
   ContainerComponent,
   DisplayObjectEvent,
+  Event,
   Event as TulipEvent,
   EventMode,
   global,
@@ -65,8 +66,48 @@ export const privateRoomComponent: ContainerComponent = () => {
     sortableChildren: true,
     eventMode: EventMode.STATIC,
   });
+
   const margin = 100;
-  allowCameraPanning($roomScene, margin);
+  let isDragging = false;
+  let dragStart = { x: 0, y: 0 };
+  let containerStart = { x: 0, y: 0 };
+
+  global.events.on(Event.POINTER_DOWN, (event) => {
+    if (event.button !== 0) return;
+    isDragging = true;
+    dragStart = { x: event.x, y: event.y };
+
+    containerStart = $roomScene.getGlobalPosition();
+  });
+
+  global.events.on(Event.POINTER_MOVE, (event) => {
+    if (!isDragging) return;
+
+    const scale = global.window.getScale();
+    const deltaX = (event.clientX - dragStart.x) / scale;
+    const deltaY = (event.clientY - dragStart.y) / scale;
+
+    let newX = Math.floor(containerStart.x + deltaX);
+    let newY = Math.floor(containerStart.y + deltaY);
+
+    const { width, height } = $roomScene.getBounds();
+
+    const minX = -width / 2 + margin;
+    const maxX = width / 2 - margin;
+    const minY = -height / 2 + margin;
+    const maxY = height / 2 - margin;
+
+    $roomScene.setPosition({
+      x: Math.max(minX, Math.min(newX, maxX)),
+      y: Math.max(minY, Math.min(newY, maxY)),
+    });
+  });
+
+  global.events.on(Event.POINTER_UP, (event) => {
+    if (event.button !== 0) return;
+    isDragging = false;
+  });
+
   $container.add($roomScene);
 
   const loadRoomPosition = () => {
